[ConfigledgerObject] lz4 compression for entry storage
