+++ src/casinocoin/unity/app_misc.cpp
@@ -34,8 +34,8 @@
 
 #include <casinocoin/app/misc/impl/AccountTxPaging.cpp>
 #include <casinocoin/app/misc/impl/AmendmentTable.cpp>
-#include <casinocoin/app/misc/impl/VotableConfiguration.cpp>
-#include <casinocoin/app/misc/impl/ConfigObjectEntry.cpp>
+#include <casinocoin/app/misc/impl/configuration/VotableConfiguration.cpp>
+#include <casinocoin/app/misc/impl/configuration/ConfigObjectEntry.cpp>
 #include <casinocoin/app/misc/impl/LoadFeeTrack.cpp>
 #include <casinocoin/app/misc/impl/Manifest.cpp>
 #include <casinocoin/app/misc/impl/Transaction.cpp>
